\name{MMGOR-package}
\alias{MMGOR-package}
\alias{MMGOR}
\docType{package}
\title{
\packageTitle{MMGOR}
}
\description{
\packageDescription{MMGOR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MMGOR}
\packageIndices{MMGOR}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MMGOR}

Maintainer: \packageMaintainer{MMGOR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
